unified-service/requirements.txt
Flask
PyMuPDF
python-dotenv
google-generativeai
